activated cross_field_diffusion_tests
