+++ include/cross_field_diffusion_broker.h
@@ -1,0 +1,12 @@
+#include "array.h"
+
+class cross_field_diffusion_broker
+{
+  public:
+
+  cross_field_diffusion_broker();
+
+  double run();
+
+  double run_1();
+};
